server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "WebSocket server"
authors = ["eilya"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
jinja2 = "^3.1.3"
openai = "^1.81.0"
google-generativeai = "^0.8.5"
python-dotenv = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
